refactor(annotations): deepen the AnnotationSyncTarget seam
Split AnnotationSyncController's three co-mingled concerns behind the
AnnotationSyncTarget seam so merge, live-pull, and push bugs each concentrate
in one place:

- AnnotationMergeOrchestrator: open-book read → merge → upsert (ADR 0038
  tombstone sweep + stale-orphan guard). Reused by the live-sync loop when a
  peer file is present.
- AnnotationLiveSync: per-book pull loop. Delegates the merge branch to the
  orchestrator; solo-namespace ticks skip the redundant re-read of our own file.
- AnnotationPushCoordinator: debounced push + close-flush, running behind the
  lock port so the durable AnnotationSweep cannot interleave.
- AnnotationLockPort: interface for the per-book annotation reconcile lock.
  ReconcileLocks implements it; tests can now stand in a fake instead of the
  process-wide singleton (DIP fix).

AnnotationSyncController becomes a ~130-line facade with the same public API
and constructor signature, so DI and existing tests remain unchanged. All
existing behavior — cross-device identity axes, ADR 0038 aged-tomb DELETE
ordering, LWW merge, PROPFIND-only ticks, sweep enqueue on push failure — is
preserved.

Closes #378
